[refactor] #133 Keychain Live 분리
